pyproject.toml
[project]
name = "ctf-pcap-generator"
version = "0.1.0"
description = "Generate realistic, flag-embedded PCAP files for CTF competitions"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1,<4",
    "gunicorn>=25,<26",
    "pydantic>=2.12,<3",
    "pyyaml>=6.0,<7",
    "requests>=2.32,<3",
    "scapy>=2.7,<3",
    "structlog>=25,<26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "ruff>=0.15",
]

[dependency-groups]
dev = [
    "pytest>=8,<9",
    "ruff>=0.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
